+++ cookbooks/aws-parallelcluster-install/recipes/base.rb
@@ -26,6 +26,7 @@ include_recipe "aws-parallelcluster-install::directories"
 
 install_packages 'Install OS and extra packages'
 
+include_recipe "aws-parallelcluster-install::base_isolated"
 include_recipe "aws-parallelcluster-install::python"
 include_recipe "aws-parallelcluster-install::cfn_bootstrap"
 include_recipe 'aws-parallelcluster-install::node'